Don't insert Error_Settings__c if in test
This is done to avoid duplicate id errors on setupOwnerId.  When running in a test with seeAllData=false, getOrgDefaults returns an in memory Error_Settings__c object even if one exists in the org.  This is the expected behavior.  However, if that same test execution tries to then insert the in memory settings object, you get a duplicate id error.  Since tests destroy their data at the end anyhow, we only need to insert the new settings object if running outside a test.