Remove Number_of_Household_Members__c reference
This field being referenced in a post-deployment bundle was causing failures when testing against the beta managed package and thus would have caused issues in the installer.  We will need to address the namespacing issues with post-deployment packages at some point soon but I want to get this build working again.