Adding expanded pattern for matching test classes
The property 'cumulusci.test.namematch' is used in a SOQL LIKE query to determine which classes to run in the python test runner.

The default value for this property is '%_TEST', but this will not match classes that end in '_TEST2', of which there are some test classes with this pattern.

This change adds an override for the cumulus.test.namematch property so that the string used in the LIKE query is '%\_TEST_'.  Here, the '%' character matches zero or more of any character, and the '_' character matches a single character.  The first '_' character in the string is escaped using a backslash so that it matches a literal '_' character.

Note that the definition in the property file escapes the backslash in the value, and this is expected to be unescaped to a single backslash when the property file is read by Ant.