Modifying "Manage Allocations" button for LEX
Converting the "Manage Allocations" button to send user to a visualforce
page, rather than a custom URL.

To do this for a list button, the target visualforce page must use the
object as its standard controller and must be set up to process a list
of objects.  Correspondingly, to apply custom logic to the page, the
controller must be set up as a custom list controller, and specified as
a controller extension for the page.

While not explicit, there is an 'id' parameter passed to the visualforce
page from a related list that contains the id of the parent record on
which the related list appeared.  I have verified this in both Classic
and LEX.