Updating Account Conversion page
Adding new JobProgressLightning component to be used with SLDS styled
pages.  This includes a new method in UIIL_JobProgress_CTRL to query for
AsyncApexJobs, as well as a new class to serialize those jobs (we no
longer expose the AsyncApexJob object directly to the caller).

Updated CONV_Account_Conversion page to use lightning styles.  This page
now uses the new JobProgressLightning component