Fix for job progress component
When job progress component was loaded via a rerender target, and not
when the page first loaded, it was failing to initialize because it was
missing the 'DOMContentLoaded' event.  Now the component checks to see
if the page has been loaded, and if so will initialize.  Otherwise, it
will initialize when the document loaded event is fired.