Job Progress Lightning component update for multiple instance
Previously, the job progress lightning component used ids to identify
the html elements that it needed to target.  However, having multiple
instances of the job progress component on a single page would cause
confusion, because the different components would access just the first
element with that id they found in the page.  This update modifies the
markup of the component to be wrapped in an apex:outputPanel tag, which
allows for a unique id to be generated for that element.  The elements
needing to be targeted are identified by querying by a data attribute as
children of the wrapper element.