Duplicate Contact Role Health Check
Adding a health check to detect the case where an opportunities has the
same contact associated via more than one OpportunityContactRole.

The reason this health check is being added is because we are changing
the behavior of how partial soft credits are counted.  In the future, we
will only credit a contact associated with an opportunity once for that
opportunity, NOT multiple times based on how many times the contact is
associated with the opportunity.

This health check brings with it some labels for displaying error,
success, and solution messages.