Update Record Type Settings updates
Move install script code that calls UTIL_RecordTypeSettingsUpdate
outside of try/catch block, so that any exceptions thrown during the
conversion process will cause the installation to fail.  Added a return
statement in the catch block, so that if any exceptions occur at an
earlier point in the install process, they are still caught, and handled
by sending an email, but will now prevent later install steps from
executing.

Mark UTIL_RecordTypeSettingsUpdate class, getInstance(), and
updateSettings() methods as globals so that they can be called from
outside of the managed package (i.e. a user can run the update script
from the Developer Console)

Switching from using the settings getInstance() and getList() methods to
querying the settings sObjects directly.  This is because getInstance()
would have returned a setting object appropriate for the running user,
but we really want to update all stored settings (including those at the
org and profile default level)

Removing some unit tests that relied too heavily on internal
implementation details, and now stub out only the minimal set of methods
to be able to perform a more end-to-end test.

Update version number checks with the expectation that this code will be
included in v3.79 of Cumulus, instead of v3.78