Fix tests that refer to record type ids
As part of the conversion from Record Type Names to Record Type Ids in
Custom Settings, many of the tests were updated to expect to receive an
Id when requesting which record type to use in that test, where
previously the test was expecting to receive a string.

There were a few spots where, if there were no available record types
for the test to use, the new code was returning null (as no Id), whereas
the old code was returning '' (as no Name).  I have fixed the places
where the code was comparing to the empty string instead of comparing to
null.

I have also updated one of the tests to avoid attempting to look up a
record type name to use in the test, then converting that record type
name to an id. This was failing when the record type name was an empty
string, similar to the above issue.  Now this test simply looks up a
record type id to use in the test, and avoids the name lookup entirely.

These issues were only evident when the user running the tests did not
have any access to any record types.

I have verified that all tests now pass if the user running the tests
has access to all record types, as well as if the user has access to no
record types.

(Any other configurations is another story....)