Make update record type settings idempotent
If the code to update the record type settings is run twice, for any
reason, then the outcome will be that all the settings will have been
cleared.

This would happen because after the first round of updating, all the
settings would be converted to and stored as ids, and so would no longer
be valid record type *names* for that org.

In order to protect against this possibility, this change makes the
update settings code first check if the value(s) stored in the setting
is already an id, and if it is then it will simply store it back to the
setting.  If it is not an id, it will attempt to look it up as a record
type name.

This should have the effect that the first time the update code runs,
it will convert all record type names to ids and store them in the
settings.  Subsequent executions of the record type update code will
read those settings as ids already, and store them back to the database
unaltered.