Date format is being saved incorrectly. -when refreshing grid format displays correctly but it is not
Date is not being edited on data import.

When you type a string in a date field and hit enter, the enter does not moves the selection to the next row, but keeps creating spaces on the field. You can't type in any other field and you need to reload the grid in order to use it again.