This adds a new Package_Settings__c custom setting object with a Namespace_Prefix__c field, and uses it to get the namespace for use in Cumulus' own weblink urls. The goal is to have buttons linking to VF pages that work in both managed and unmanaged deployments of NPSP.
For managed NPSP, the setting is initialized in the STG_InstallScript after installation or push upgrades.
For unmanaged NPSP, the setting is initialized in the anon apex run by the npsp_default_settings CCI task.

The new setting object is set as Protected because to use it from outside the package you would need to already know the namespace.

We should be sure to test this as a beta release to make sure there are no surprises.