Populate relationship from obj mapping and PR changes
Added the population of the relationship field on child additional objects from the object mapping rather than explicitly through field mappings.  Clients still have the choice to use it or not if they use custom object mapping logic.  This entailed removing a the OCR field mappings for opportunity id since it will now be populated automatically by the relationship on the object mapping.

Also made a wide variety of changes in response to feedback from pull request, mostly name changes and debug removal.