Removing Campaign field mapping from GAU Allocation
Removing the Campaign field mapping from the GAU Allocation Object Mappings because it was causing issues with detecting whether a new GAU was actually trying to be created or whether there was just a campaign specified.  Since there is no object mapping for the campaign it was not possible to exclude the Campaign field mapping using the non-relationship field logic for determining whether to create a new record.  The campaign lookup wouldn't actually be used from BDI anyway, so this just simplifies things as well. 

Also making the GAU allocation percent not required since that can interfere with payment allocations.