Update the BeforeInsert trigger on Contact to apply the Contact's address to the Account record when it's being created. The previous behavior was to leave the Account address fields blank, then allow the Contact's trigger sets to create the Address and then apply that Address back to the Account. This caused two additional sets of DML operations. None of this was really necessary if the Address is applied to the Account object when it's created.
However, it appears that fixing the extra DML in the address logic when inserting a Contact with an Address when the AccountId field is null now allows the duplicate rules to actually run when inserting Account records. Before this change there must have been something in the sequence that caused the active dupe rule to be ignored. To fix this, two changes were made to three test methods: use the dmlHeader option to allow dupes or make sure the lastname+address combo is unique on the Contact records being inserted.