Grouping SObjects to reduce DML
This change groups together different SObject types into 1 update or insert DML statement to reduce the number of DML statements.  It also accounts for the fact that there are a max of 10 different SObject types allowed in a single dml statement and will break SObjects into multiple lists as needed.