Data Import Re-processing: Skip already created Donations
With Advanced Mapping enabled, it is possible for a Data Import record
to create a Donation successfully but then fail when processing
additional objects.  This change skips over any further processing of a
Donation that is in "Created" status, allowing a re-processing attempt
of the additional objects on subsequent Data Import processing runs.