Adding filtering of the Link To Record picklist by object
Adding filtering to the picklist that is used for the "Link to Record" field on the object mapping modal. It will now only allow "Reference" fields that actually look up to the object defined in the Object Name field.  It will still allow Text fields for selection to support those standard objects that cannot be looked up to.