Adding ability to define one field for matching for additional objects
For the BDI Additional objects functionality this allows a user to define one field mapping for use in matching per Object Group.  This also includes creating an overrideable class BDI_ObjectMatchingLogic, which contains the logic for matching and in the future will be able to be replaced by client's custom logic.

In addition, this moves the log error functionality into the object wrapper in order to streamline the code and enable errors to be attached from outside of the Additional Object Service.