Fix of Unit Test failures
Because of the distance check in the base class, and because we were using an arbitrarily low batch size with a higher chunk size, it was intermittently running in LDV mode and erroring because it's batch size was 1 but it's chunk size was 10, which meant there would be multiple batches returned for a single batch execution, which is a nono...this is also why the other tests, and the ldv test were not failing, and why the testing of the implementing batch went fine. It will now use a default chunk and non ldv sizes, with a larger batch size so this can never occur.  Basically, I picked a chunk and max non ldv size that was small enough that it would decide to run in ldv mode when it was not expected to...and then using settings that caused it to break.