Convert widget data to BDI_DynamicSourceGroup and store in JSON on DI
Contains two different implementations of the conversion of widget data to the BDI_DynamicSourceGroup class format, which will be serialized into the Additional_Object_JSON__c field on the DataImport__c object.

One of the implementations assumes that the objects in the widget data map are dummy DataImport__c objects with just the appropriate fields populated for the template object mapping.

The other implementation assumes the objects in the widget data map are the target object, and then reverse maps them into dummy DataImport__c objects for processing by BDI.