Removing case check on relationship field and tweaking metadata
Removing the case check on relationship since the GAU allocation object mappings have a relationship field that doesn't contain the npsp__ namespace prefix, which then causes an error in namespaced orgs when checking the case.  Since we can't easily fix this in all existing customer orgs that are currently using Advanced Mapping, we are removing the check for now. 

Also, I'm checking in a change to add those namespaces to make sure future orgs have the correctly namespaced field name.