Enable edit of batch information in modal using geBatchWizard
In order to display the batch create/edit wizard in a modal, the
new GE_GiftEntryForm (Aura) component wraps the geBatchGiftEntryApp
(LWC) component so that the lightning:overlayLibrary can be used to
provide modals needed by the LWC app.  geFormRenderer is then removed
from geHome and display of the form is handled by the new Aura
component.