Update app level storage of selected donor values
Account1Imported and Contact1Imported are special case fields as they
are the only two that represent a donor.  These changes store local
variables at the App-level to track the current value for each of these
fields by handling events from the Lookup field components.  The local
App-level vars are then referenced to determine how to set the values
for the Review Donations component when the Donation Donor field is set
and reduces the amount of looping/inspecting of the JSON template data
structure that was being done before to get the values of these fields.