Use getRecord to get displayValue when only an Id is passed in
These changes are intended to have the geFormFieldLookup.js cmp use its
getRecord function to get the display value when only an Id is passed
in, making it easier to load sibling fields when a selected record has
sibling lookup values to load.  If the cmp is passed a display value
along with an Id, it will use that and getRecord should not be called.