Update on CON_ContactMerge_CTRL for DRS_Contact_Merge
Added two methods to this class:
redirectToShowDRSList() - This method is called on 'Show Duplicate Record Sets' button click. It will redirect to the DRS_ContactMerge page.
showContactSearchPage() -This method is called on 'Search by Contact' button click. This method is to reinitialise the variables used on  search by contact page.