Add reset function to geFormFieldPicklist
The reset function sets the picklists recordTypeId back to its default
so that the component will set its options back to the default options
set, and the geFormField fires the handlePicklistChange event after
resetting to communicate its new value upward in the dom.

This change should fix an issue where re-selecting the same donor after
cancelling the form would not re-display the Review Donations link.