Fix access check failure when closing modal in debug mode
With debug mode enabled, a Component Error with the message:

Access Check Failed! AttributeSet.get():
attribute 'isLoading' of component ...

would be thrown when the Review Donations modal closed.  This change
fixes that debug-mode-only error by removing the function that
references the isLoading property. This should be ok because the
isLoading property appears to be unused with this modal
and is always undefined when it closes.