Created Contact Merge Page Settings custom setting
Since we were facing governor limit issues(too many query rows) due to the count of DRS, DRI and related contact records, we have kept the DRS query limit count configurable in this custom setting. For orgs with large set of DRS records, we have recommended in the help text the count to be kept below 7k. Default count would be 5k if not set.