Update DRS_ContactMerge_CTRL as per review comment
Renamed DRS_ContactMerge_CTRL to CON_DRSContactMerge_CTRL
Moved the common contact merging logic in the CON_ContactMerge_UTIL utility class
Since we were facing governor limit issues(too many query rows) due to the count of DRS, DRI and related contact records, we have kept the DRS query limit count configurable in a custom setting. For orgs with large set of DRS records, we have recommended in the help text the count to be kept below 7000. Default count would be 5000 if not set in the custom setting.