Get RecordType Name from stored objectInfo object instead of querying
When a records fields are loaded into the form in response to a lookup
being populated, if only the RecordType Id is included these changes
get the RecordType Name from the objectInfo object already stored inside
this component, which prevents the geFormFieldLookup component from
having to query for the Name.

This commit also includes additional commenting to provide clarity
around the special case handling for RecordTypeId fields.