Fixing help text mapping migration with Payment_Type__c
An error was being displayed for the new Payment_Type__c field on DataImport__c during helptext mapping to custom metadata migration.  This was due to the way that the migration logic was identifying fields with help text mappings that needed to be moved. 

In the short term this commit adds the help text of the payment type field to the list of fields to ignore.

 In the long run the logic in the BDI_MigrationMappingUtility in HelpTextFieldMapping that deals with identifying if there are dataImportObjects needs to be changed (at least the 'containsAny' line, but that will need to be tested more carefully to make sure it is still catching field mappings that should have been migrated, but failed to migrate.

This commit is also adding package.xml updates for the new Data_Import_Field_Mappings.