Add test method validating required config present for Elevate customers
This set of changes defines a list of configuration key-value pairs that
must be present in an org for it to be considered an Elevate customer.

The elevateSDK and productId keys are required to load the Elevate
SDK into the iFrame.

The baseURL, apiKey, sfdo.merchantids, and sfdo.gatewayids keys are
required to make callouts to the Elevate endpoints.