Add aura enabled method to check elevate customer status
This commit changes the isElevateCustomer method to a property on
the GE_PaymentServices class and introduces an aura enabled method
on the GE_GiftEntryController so that exceptions can be wrapped in
AuraHandledExceptions there and to route apex requests for GE through
the dedicated GE_GiftEntryController class.

The GE_PaymentServices.isElevateCustomer property will query and read
configuration parameters from the database each time it is read -
clients should handle caching the results if/as appropriate for their
use case.