Do not hide GAU Allocations widget when cannot retrieve Pmt Serv Config
Prior to this commit, if the retrieval of the Payment Services config
record threw an exception, the the GAU Allocations widget would also be
hidden from the template builder sidebar.  This change fixes that so
that the GAU Allocations widget will still show up in the sidebar if
there is a problem retrieving the Payment Services configuration
records and only the Salesforce.org Elevate widget will be hidden
in that case.