Decouple retrieval of namespaceWrapper from retrieval of field mappings
The retrieval of the field mappings can be deferred but to prepare for
that change the retrieval of the namespace wrapper is decoupled since
the list view needs the namespace initially to construct record links.

This also changes the retrieval mechanism to import the asset into
whichever component needs it using a cacheable apex method rather than
storing it in a variable in a utility component.