Gift Entry: Centralize state management
Introduce a formState variable in the geFormRenderer component that
will hold the state of the form and pass that state down to child
components that will read their values from it.  Introduces the
formfieldchange event that will send all geFormField changed values
up to the renderer which will update its formState object, allowing
other sibling components to update their UIs accordingly via reactivity.

For this WI, picklists will use the formState object to read their
values and react when their value is changed or when their record type
is changed, via either lookups being selected or record types being
included on the form.  In future WIs other field types will be updated
to also read their values from the formState object.

This WI also initializes the formState object with both default values
and/or values from the Account or Contact which was the entry point by
reading values that currently get embedded into the form template.