Do not fire lookuprecordselect on every render to prevent
the querying of selected lookup records when loading from a row in the
batch gifts table.  The open row action should load the saved version of
the DataImport record to preserve any updated values the user may have
entered and should not reload selected lookups from the database to
prevent overwriting user-modified values.