Do not run reset flow when opening a row from the batch gifts table
the reset flow will requery the database for any lookups that have
values which could overwrite any user-modified values.  This change
promotes opening the row from the gifts table as it exists in the table
at the time of opening with no side effects or modifying of the rows
data.