Add the "User Selected Match" status to the DI when DonationImported
selected manually.  Previously, when a User selected a value for the
donation imported field manually, the lookup value would not be saved
even though the selected Opps related field values were loaded and saved
.  This change sets the donation import status field to user selected
match, which causes BDI to retain the selected lookup value so that
the selected opp stays selected after dry run and when opening the
row subsequently from the gifts table.