Preserve fields with blank values when submitting to server for save
SaveAndDryRun does not return fields that had null, undefined, or empty
string values when submitted.  This change preserves those fields and
applies them back to the response object before upserting into the gifts
table so that when they are opened, the values (which might have been
user-input) are preserved and are not overwritten by defaults.