When loading a selected record's fields from db, apply defaults if the
records field has a blank value.  This is necessary because the
reset function inside geFormField used to automatically apply the
default value when null was passed in via the load flow.  Now, the
renderer applies default values to selected records for blank fields,
and keeps database values where fields have values in the db.  This
allows passing in null and NOT applying default values when opening
from a row in the table, but still applies defaults to otherwise
blank fields when selecting lookup records.