Reset fields in formState to null before initializing to their default
initial values.  This prevents fields that receive a value after getting
a selected record from retaining that value when the fields are
subsequently passed to "set initial value".

Before this change, fields that did not have a default value would
retain their state from a previous record selection after calling
"setInitialValueInFormStateForFieldMappings" because their initial
value is undefined, which does not overwrite an existing value
in formState when calling "updateFormState".