Only update formState with field values that are in use in the template
When lookups are populated, all mapped fields on that object are queried
and populated in formState.  This change only populates formState with
field values that are in use in the Template, in order to prevent saving
field values to the Data Import record that are not in use on the form.