W-8320450 - POC2 for RD GAU Allocations in BDI - using AllocService
A proof of concept for implementing GAU Allocations for RD In BDI by disabling all allocation triggers before BDI Processing, then using the existing additional objects code for inserting GAU Allocations, then running ALLO_AllocationsService manually and re-enabling allocations triggers.

If an error is encountered during allocations processing then the allocations are manually deleted and the error is set in one of the import status fields.  The downside of this approach is that any client triggers that rely on allocations would presumably fire even on the insert of an invalid allocation (such as percent > 100%) since it wouldn't be caught before insert.