[wip] one last time
There's some weirdness in the github context that I don't understand
very well. By all acounts, $GITHUB_REF is pointing at a something we
don't want. Here I try to checkout the ref given by the PR, and assign
it to an environment variable for good measure.

If this won't work...we may need an action (e.g. octokit, etc).