Remove flag tracking number of times HH Naming is run
This flag was introduced to limit the number of times this method runs unnecessarily, but with Gem installed, it needs to run more that the allotted number of times due to how HH Naming works for LegacyHouseholds. This commit allows the fix to be deployed via 2gp and get insight into this methods behavior.