W-10034496 - Fix Dupe Oppts when creating RD with camp in BDI
This fixes an issue where if a Recurring Donation is created in BDI with a campaign defined, but no initial Donation defined it still creates an initial donation.  This was occurring because the code that checks to see if any Opportunity Object mapping fields were populated was including the DonationCampaignImported__c field (Which is used by both RD and Oppts) in the populated field check that determines whether or not the Opportunity attempts to be created.