W-10034496 - Add Donation Campaign Source to default mappings
This field mapping was missed when we originally converted the help text mappings to the custom metadata "advanced" mappings.  It has been getting created automatically during advanced mapping enablement so it hasn't directly caused problems, but the fact that it is missing meant that the unit tests for Recurring Donation didn't pick up on the fact that an extra unnecessary null donation gets created if you are creating a Recurring Donation (and no initial donation) and include a campaign.  Now that it is in there, the unit tests should be able to pick up any regression on this bug.