Do not manually set the state code when copying address values onto SObject records
The in memory map validStatesByLabel cannot correctly determine which value to use when there is more than one state/province in the system with the same name. The platform does look at the country and our country code to determine which state code values are valid. This change allows the platform to set the state code value.