Do not populate state or country using code values
If either the code or the picklist String field are populated the platform automatically populates the other value. Which means if there is a state or country code present then they're also will be a value in the picklist field. So the above mapping to state and country fields should be sufficient.