Remove in memory maps, allowing platform to populate the state code and or the state based on the value in the other field.
This allows the platform to handle state and state code, country and country code values as intended with the state country code picklist feature.

Also, Check to see whether state and country picklist are enabled by checking for billing state code presents on the account object. Previously the check would initialize the in memory maps. Which are no longer being used.