W-10707489 - Ignore existing Default GAU Allocations when processing new GAU Allocations in BDI
Currently in BDI if you have a DI with an existing Oppt or Payment entered that has an existing GAU Allocation linked to it, then BDI will intentionally throw an error since it cannot properly resize existing GAUs.  This doesn't need to apply to Default GAU Allocations though since those can be automatically resized.

This ticket basically excludes the default GAU from the check for existing GAU Allocations since it is not relevant and will automatically resize if needed.  This will prevent the Error from being thrown in a large number of cases in an org with Default GAU enabled.

Note: This ticket doesn't fix this issue if 'Payment Allocations' are enabled because in that scenario only the Opportunity Allocations are updated, and the payment allocations become out of sync with the Opportunity Allocations.