Fix unit test error if there are no active contact duplicate rules
Fixes issue where a unit test would throw an exception if the org it is being run in has no active contact duplicate rules.

Unfortunately right now there doesn't appear to be any way to check if there are active duplicate rules, so we just have to rely on the findDuplicates method throwing an error.  This can be improved if/when it becomes possible to check on what rules are active.