W-12250612 - Fixed issue with BDI matching when using Dupe Mgmt
Fixed incorrect internal matches with BDI matching when using the 'Contact Duplicate Rules' (aka Duplicate Management) option.  Essentially if you had two Data Import records with any one of the 4 pieces of contact information used for matching (first name, last name, email, phone) matching, then it would match the two records together incorrectly and only create one new contact.  It should only match internally like that if all the pieces of contact information match.

The cause of the issue was that because the Contact matching rule name didn't include the name of any of the fields used for matching, it was creating a key for matching for each field with just the field populated and all others blank.  This meant that if any of these contact fields matched, the whole contact would be matched (which is very problematic for things like first name).

The fix was to basically make sure that if Duplicate Management was enabled, then it should never include a blank value unless the contact field was actually blank.

Also updated the test class to test the partial contact match scenario and the exact internal duplicate match scenario.  Also added tests for non-dry run scenarios.